Changes:  - updated msvc/config.jam  - workarounds for compiler warnings
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@5288 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ tools/gltsgraph/utils.cpp
@@ -1,7 +1,7 @@
 #include "utils.h"
 #include <math.h>
 
-namespace Utils 
+namespace Utils
 {
 
 Utils::Vect operator+(Vect v1, Vect v2)
@@ -32,7 +32,7 @@ Utils::Vect operator/(Vect v, double s)
 
 float vecLength(Vect v)
 {
-  return sqrt(v.x * v.x + v.y * v.y);
+  return static_cast< float >(sqrt(v.x * v.x + v.y * v.y));
 }
 
 float angDiff(Vect v1, Vect v2)
@@ -45,7 +45,7 @@ float angDiff(Vect v1, Vect v2)
 }
 
 float signum(float v)
-{ 
+{
   if (v < 0)
   {
     return -1.0f;
